chapter-3/requirements.txt
gptcache==0.1.44
langchain==0.2.12
chromadb
langchain_chroma==0.1.2
langchain_community==0.2.11
langchain_openai==0.1.20
openai==1.40.1
python-dotenv==1.0.1
onnxruntime
unstructured
"unstructured[all-docs]"
python-magic-bin
